Server/requirements.txt
fastapi[standard]
uvicorn
websockets
pydantic
pydantic-core
python-multipart
google-genai
python-dotenv
redis
fakeredis
posthog
google-generativeai
sentry-sdk[fastapi]